vps/requirements_vps.txt
python-telegram-bot==22.7
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiofiles
httpx
